Refactor and rename TaskApi components
The Task API elements have been refactored, with several files being removed and others being renamed for clarification. The primary changes involve renaming `TaskHubSpotModel` to `HubSpotTaskModel` and `TaskListHubSpotModel` to `HubSpotTaskListModel`. The `IHubSpotTaskApi` interface has been changed to `IHubSpotTasksApi` with the associated implementation class renamed to match. This also involved updates to related methods and properties in other classes.
